requirements.txt
python-dotenv >=0.9.1
selenium >=3.141.0